+++ 1주차/조민의/Main_Hash_001.java
@@ -3,7 +3,7 @@ import java.util.Iterator;
 import java.util.Map;
 
 
-public class Hash {
+public class Main_Hash_001 {
 	public String solution(String[] participant, String[] completion) {
 		String answer = ""; //초기화 값
 		HashMap<String, Integer> map = new HashMap<>();//key, value의 값을 해쉬로 나타냄
@@ -25,6 +25,27 @@ public class Hash {
 		Solution_Sort sol = new Solution_Sort();//완주하지못한 사람을 담을 배열
 		System.out.println(sol.solution(part, comp)); //완주 못하는 사람이 나오는 식
 		}
+
+
+public class Solution_Sort {
+	public String solution(String[] participant, String[] completion) {
+		// 1. 두 배열을 정렬한다
+		Arrays.sort(participant); 
+		Arrays.sort(completion);
+		// 2. 두 배열이 다를 때까지 찾는다 
+		int i = 0; for(i=0;i<completion.length;i++) 
+			if(!participant[i].equals(completion[i])) break;
+		// 3. 여기까지 왔다는 것은 마지막 주자가 완주하지 못했다는 의미이다.
+		return participant[i];
+		} 
+	public static void main(String[] args){
+		String[] part = {"leo", "kiki", "eden"};
+		String[] comp = {"eden", "kiki"};
+		Solution_Sort sol = new Solution_Sort();
+		
+		System.out.println(sol.solution(part, comp)); 
+		}
+	}
